\name{kFcheck}
\alias{kFcheck}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Visualisation of isopter values
}
\description{
Plots curves or points for multiple individuals on a kinetic grid
}
\usage{
kFcheck(mat.output, name.iso, plot.lines = TRUE, title1 = " ")
}
%- maybe also "usage" for other objects documented here.
\arguments{
  \item{mat.output}{data.frame generated by \code{kf.sort}}
  \item{name.iso}{character, one of "III4e", "I4e" or "I2e"}
  \item{plot.lines}{
logical, if \code{TRUE} individual isopters are plotted, otherwise individual data points are displayed}
  \item{title1}{character, specifying plot title}
}

\value{
Graphical output
}

\author{
Dipesh E Patel & Mario Cortina-Borja
}

\examples{
## Only runs after kf.sort has run
test3<- kf.sort()
kFcheck(test3$mat.output, "III4e", title1="III4e data", plot.lines=FALSE)
}
