\name{kineticF-package}
\alias{kineticF-package}
\alias{kineticF}
\docType{package}
\title{
Framework for the Analysis of Kinetic Visual Field Data
}
\description{
A collection of functions covering data cleaning, processing,
 visualisation and analysis for manual (Goldmann) and automated (Octopus 900) kinetic visual field data. 
The analysis is primarily aimed at summarising normative data, with code  provided
 to allow programmers to adapt the basic functions to their specific needs.
}
\details{
\tabular{ll}{
Package: \tab kineticF\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-06-04\cr
License: \tab GPL (>= 2)\cr
}}
\author{Dipesh E Patel <dipesh_patel@ucl.ac.uk> and Mario Cortina-Borja <m.cortina@ucl.ac.uk>}



\keyword{ kineticF }


