\name{set.template}
\alias{set.template}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Kinetic visual field template
}
\description{
Opens a plot window and displays a kinetic perimetry grid
}
\usage{
set.template(void = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{void}{Adds the 'void' areas of a Goldmann field to the kinetic plot. Default is \code{TRUE}.}
}

\value{
Graphical output
}

\author{
Dipesh E Patel & Mario Cortina-Borja
}

\examples{
set.template(void=FALSE)
}

