\name{stop.identify}
\alias{stop.identify}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Stops the process of re-ordering a matrix of coordinates
}
\description{
Changes the order in which a matrix of coordinates is plotted to allow closure on a polygon.
This function is for internal use and is not meant to be called by the user.
}
\usage{
stop.identify(xy)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{xy}{matrix of coordinates}
}

\value{
A re-ordered matrix of coordinates
}

\author{
Dipesh E Patel & Mario Cortina-Borja
}
